Implemented VideoClips to load video files
The new class VideoClips is organized into the loki/processing directory 
for processing data files. VideoClips takes a list of video files, loads 
each file using moviepy, and then stores it in a list. It can then write 
out sections of the video files with options for file format, fps, and 
file names.