Implemented predict method in VolumeModel
The VolumeModel.predict() is now implemented. It will search over a 
given set of clips and output the video index and time-indices for the 
portion of the video with the highest average volume. Currently, it will 
check if an audioclip is shorter than the search window. If so, it then 
computes the average volume of the whole clip.