Refactor: Use pytorch not Keras and Tensorflow
Turns out there are serious issues when trying to get tensorflow to play 
nicely with a pretrained model and layers added on. Switch to using 
Pytorch so that the code can still be made modular for now. Keep the 
tensorflow pretrained model and continue using it as a modular piece in 
the processing step.