Implement function for stacking embeddings
For a small network, every second is trained on in sequence. This 
function takes a list of embeddings (i.e. each embedding is 10x128) and 
stacks them along the zeroth axis so that you have a single Nx128 array. 
If targets/classifications were given for each clip, then targets is 
also extended so that every second gets the same class as the whole 
clip.